refactor: 모달 폴더 구조 변경 photoList/guest, photoList/host
