+++ frontend/src/components/specific/modal/photoList/guest/GuestPhotoListModal.styles.ts
