+++ frontend/src/components/specific/modal/photoList/host/HostPhotoListModal.styles.ts
